Use native HTTP APIs instead of octokit
octokit is quite a heavy dependency. I used it because I didn't want to
implement this part of tool from scratch (I was lazy!), but Next PR
Number only uses it for one kind of HTTP requests. Other than
substituting the owner and name into the URL, it's fixed. Using native
APIs from the browser directly will take some work, but it should reduce
page load times significantly, enough to be worth it.

Now Next PR Number is using a XMLHTMLRequest based approche. Since I
stil felt lazy, I add a much of (copied and pasted) boilerplate so I
can continue to use the same request management logic. Wrapping it in a
promise is definitely not necessary but it's fine :p